修改bufferpool单位为long

+++ src/main/java/io/mycat/net/BufferQueue.java
@@ -10,10 +10,10 @@ import java.util.LinkedList;
  * @author wuzh
  */
 public final class BufferQueue {
-	private final int total;
+	private final long total;
 	private final LinkedList<ByteBuffer> items = new LinkedList<ByteBuffer>();
 
-	public BufferQueue(int capacity) {
+	public BufferQueue(long capacity) {
 		this.total = capacity;
 	}
 
@@ -22,11 +22,11 @@ public final class BufferQueue {
 	 * 
 	 * @return
 	 */
-	public int snapshotSize() {
+	public long snapshotSize() {
 		return this.items.size();
 	}
 
-	public Collection<ByteBuffer> removeItems(int count) {
+	public Collection<ByteBuffer> removeItems(long count) {
 
 		ArrayList<ByteBuffer> removed = new ArrayList<ByteBuffer>();
 		Iterator<ByteBuffer> itor = items.iterator();
